requirements.txt
mcdreforged>=2.13.0
aiocqhttp>=1.4.4
aiohttp>=3.8.0
pyyaml>=6.0.1
python-telegram-bot>=21.10